Add auth router
