+++ src/routes/index.js
@@ -1,2 +1,3 @@
 export * from './trails';
 export * from './newUser';
+export * from './auth';
